test(workbench): run the readFinalizedEvalRun fake-client tests on the unit pool
The three finalization-polling tests exercise readFinalizedEvalRun with
in-memory fake clients and never touch a browser or server, yet they
lived in evals-real.e2e.test.ts and paid the serialized integration
pool for it. They move verbatim to evals-finalized-run.test.ts, which
the unit config picks up by default.
