+++ packages/workbench/tests/evals-real.e2e.test.ts
@@ -12,7 +12,6 @@ import { createWorkbenchAssetSource } from '../../agent-bundle/src/dev/workbench
 import { startDevServer } from '../../agent-bundle/src/dev/workbench-server.ts';
 import { createProjectFixture, removeProjectFixture } from '../../agent-bundle/tests/helpers/project-fixture.ts';
 import { seedEvalProject, writeEvalSuite } from '../../agent-bundle/tests/support/eval-project.ts';
-import { readFinalizedEvalRun } from '../src/evals/evals-page.tsx';
 import { closeServer } from './support/http.ts';
 import { workbenchBrowserAliases } from './support/workbench-browser-modules.ts';
 import { buildWorkbench, e2e, workbenchAssets, workspaceRoot, workbenchUrl } from './support/workbench-e2e.ts';
@@ -20,55 +19,6 @@ import { buildWorkbench, e2e, workbenchAssets, workspaceRoot, workbenchUrl } fro
 const evalsPage = join(workspaceRoot, 'packages', 'workbench', 'src', 'evals', 'evals-page.tsx');
 const browserTimeout = 12_000;
 const runCompletionTimeout = 60_000;
-
-e2e('retries a terminal canonical read until the durable run finalization is visible', async () => {
-  let reads = 0;
-  const waits: number[] = [];
-  const result = await readFinalizedEvalRun({
-    client: {
-      read: async () => ++reads === 1
-        ? { run: { completedAt: undefined } } as never
-        : { run: { completedAt: '2026-08-18T00:00:02.000Z' } } as never,
-    },
-    runId: 'run-terminal-race',
-    signal: new AbortController().signal,
-    wait: async (milliseconds) => { waits.push(milliseconds); },
-  });
-
-  expect(reads).toBe(2);
-  expect(waits).toHaveLength(1);
-  expect(result.run.completedAt).toBe('2026-08-18T00:00:02.000Z');
-});
-
-e2e('surfaces a terminal canonical-read error without retrying it', async () => {
-  let reads = 0;
-  let waits = 0;
-
-  await expect(readFinalizedEvalRun({
-    client: { read: async () => { reads += 1; throw new Error('invalid durable DTO'); } },
-    runId: 'run-terminal-error',
-    signal: new AbortController().signal,
-    wait: async () => { waits += 1; },
-  })).rejects.toThrow('invalid durable DTO');
-
-  expect(reads).toBe(1);
-  expect(waits).toBe(0);
-});
-
-e2e('stops bounded terminal finalization polling instead of looping forever', async () => {
-  let reads = 0;
-  let waits = 0;
-
-  await expect(readFinalizedEvalRun({
-    client: { read: async () => { reads += 1; return { run: { completedAt: undefined } } as never; } },
-    runId: 'run-terminal-timeout',
-    signal: new AbortController().signal,
-    wait: async () => { waits += 1; },
-  })).rejects.toThrow('Recorded eval results were not finalized in time.');
-
-  expect(reads).toBe(8);
-  expect(waits).toBe(7);
-});
 
 const listen = async (server: Server): Promise<string> => {
   server.listen(0, '127.0.0.1');
